02_ai_scraping_ai/requirements.txt
trafilatura
newspaper3k
feedparser
requests
google-generativeai
aiohttp
sentence-transformers
chromadb
python-dotenv
streamlit
pandas
numpy
plotly
scikit-learn
umap-learn
